Add CI to run tests with 7.6 and master (#7)
Using GitHub Actions, builds modules and runs tests (one manually specified test file). Test is on Ubuntu 18.04 with UbuntuGIS PPA (currently 7.6) and self-compiled binary from code in master (currently 7.9).

We do not test with 7.4 which is available without PPA because it is old for us spend time with reproducibility issues there (the test does not run with the same numbers).

Make test an executable file and adds missing shebang (fixes Exec format error).